Refactors token verification to send token as query param, not header
This is a code smell for me, but I'm not sure when I'll be able to get back and fix it. When using token auth header, the API returns duplicate character data for diverse tokens, and I am not sure why. I consider the current, working state to be a hack.